refactor: Move unused WASM interfaces and Rust modules to legacy
- Moved 70+ unused WASM functions to legacy/lib/wasm/unused-wasm-interfaces.ts
- Cleaned lib/wasm/index.ts from 900 to 305 lines, keeping only 12 essential functions
- Moved 5 Rust modules (bezier, physics_utils, spatial, memory, batch_transfer) to legacy/wasm/src/
- Updated wasm/src/lib.rs to only include active modules (math, star_field)
- Updated legacy/README.md with comprehensive WASM cleanup documentation

Active codebase now only includes:
- Star generation (positions, colors, sizes)
- Star effects (twinkle/sparkle with temporal coherence)
- Frustum culling (visibility optimization)
- Basic math utilities (fast trigonometric functions)

All legacy functions preserved for reference and potential restoration.
