+++ legacy/wasm/src/bezier.rs
