+++ legacy/wasm/src/memory.rs
