+++ legacy/wasm/src/spatial.rs
