+++ lib/wasm/index.ts
@@ -1,4 +1,4 @@
-// WASM module loader with graceful fallback
+// WASM module loader with graceful fallback - Active functions only
 
 let wasmModule: any = null;
 let loadPromise: Promise<void> | null = null;
@@ -7,142 +7,22 @@ let isUsingFallback = false;
 export interface WASMModule {
   add: (a: number, b: number) => number;
   greet: (name: string) => string;
-  // Star field functions
+  // Star field functions - core functionality only
   generate_star_positions: (count: number, start_index: number, min_radius: number, max_radius: number) => Float32Array;
   generate_star_colors: (count: number, start_index: number) => Float32Array;
   generate_star_sizes: (count: number, start_index: number, size_multiplier: number) => Float32Array;
-  calculate_star_effects: (positions_ptr: number, count: number, time: number) => Float32Array;
-  calculate_star_effects_into_buffers: (positions_ptr: number, twinkles_ptr: number, sparkles_ptr: number, count: number, time: number) => void;
   calculate_star_effects_arrays: (positions: Float32Array, count: number, time: number) => Float32Array;
-  calculate_twinkle_effects: (positions_ptr: number, count: number, time: number) => Float32Array;
-  calculate_sparkle_effects: (positions_ptr: number, count: number, time: number) => Float32Array;
-  calculate_rotation_delta: (base_speed_x: number, base_speed_y: number, speed_multiplier: number, delta_time: number) => Float32Array;
-  calculate_speed_multiplier: (is_moving: boolean, click_time: number, current_time: number, current_multiplier: number) => number;
+  calculate_star_effects_with_temporal_coherence: (positions: Float32Array, previous_twinkles: Float32Array, previous_sparkles: Float32Array, count: number, time: number, threshold: number) => Float32Array;
   // Camera frustum culling
-  cull_stars_by_frustum: (positions: Float32Array, count: number, camera_matrix: Float32Array, fov: number, aspect_ratio: number, near: number, far: number) => Uint8Array;
   get_visible_star_indices: (positions: Float32Array, count: number, camera_matrix: Float32Array, margin: number) => Uint32Array;
   cull_stars_by_frustum_simd?: (positions: Float32Array, count: number, camera_matrix: Float32Array, margin: number) => Uint8Array;
-  // Enhanced SIMD batch processing for LOD groups
-  calculate_star_effects_by_lod: (near_positions: Float32Array, near_count: number, medium_positions: Float32Array, medium_count: number, far_positions: Float32Array, far_count: number, time: number, quality_tier: number) => Float32Array;
-  // Temporal coherence optimization
-  calculate_star_effects_with_temporal_coherence: (positions: Float32Array, previous_twinkles: Float32Array, previous_sparkles: Float32Array, count: number, time: number, threshold: number) => Float32Array;
-  get_stars_needing_update: (positions: Float32Array, previous_twinkles: Float32Array, previous_sparkles: Float32Array, count: number, time: number, threshold: number) => Uint32Array;
-  calculate_star_effects_temporal_simd?: (positions: Float32Array, previous_twinkles: Float32Array, previous_sparkles: Float32Array, count: number, time: number, threshold: number) => Float32Array;
-  // LOD distribution calculations
-  calculate_lod_distribution: (total_count: number, quality_tier: number) => Uint32Array;
-  // Frame rate calculation
-  calculate_fps: (frame_count: number, current_time: number, last_time: number) => Float32Array;
-  // Math utilities
+  // Math utilities - essential functions only
   fast_sin: (x: number) => number;
   fast_cos: (x: number) => number;
   fast_sin_batch: (values: Float32Array) => Float32Array;
   fast_cos_batch: (values: Float32Array) => Float32Array;
   seed_random: (i: number) => number;
   seed_random_batch: (start: number, count: number) => Float32Array;
-  // Bezier functions
-  precalculate_bezier_path: (start_x: number, start_y: number, control_x: number, control_y: number, end_x: number, end_y: number, segments: number) => Float32Array;
-  precalculate_bezier_paths_batch: (paths_data: Float32Array, segments: number) => Float32Array;
-  interpolate_bezier_point: (points: Float32Array, t: number) => Float32Array;
-  precalculate_cubic_bezier_path: (p0x: number, p0y: number, p1x: number, p1y: number, p2x: number, p2y: number, p3x: number, p3y: number, segments: number) => Float32Array;
-  calculate_bezier_length: (points: Float32Array) => number;
-  // Memory management
-  SharedBuffer: any; // Constructor type
-  DirectMemory: any; // Static methods type
-  MemoryPool: any; // Constructor type
-  batch_process_sin: (input: Float32Array) => Float32Array;
-  batch_process_cos: (input: Float32Array) => Float32Array;
-  batch_process_with_operation: (input: Float32Array, operation: string) => Float32Array;
-  // Spatial indexing system
-  SpatialGrid: any; // Constructor type
-  // Particle pool and utilities
-  ParticlePool: any; // Constructor type
-  PhysicsUtils: any; // Static methods type
-  Force: any; // Constructor type
-  FastRandom: any; // Constructor type
-  BatchTransfer: any; // Static methods type
-  TypedBatchTransfer: any; // Static methods type
-  ViewTransfer: any; // Constructor type
-  NebulaSystem: any; // Constructor type
-  // Memory exports
-  __wbindgen_export_0?: WebAssembly.Memory;
-  __wbg_memory?: WebAssembly.Memory;
-  memory?: WebAssembly.Memory;
-}
-
-// Memory management classes interfaces
-export interface SharedBuffer {
-  new (size: number): SharedBuffer;
-  from_data(data: Float32Array): SharedBuffer;
-  size: number;
-  ptr: number;
-  ptr_mut(): number;
-  write(data: Float32Array, offset: number): void;
-  read(): Float32Array;
-  read_slice(start: number, length: number): Float32Array;
-  apply_sin(): void;
-  apply_cos(): void;
-  apply_operation(op: string): void;
-  free(): void;
-}
-
-export interface DirectMemory {
-  allocate_f32_array(size: number): number;
-  free_f32_array(ptr: number, size: number): void;
-  copy_from_js(ptr: number, data: Float32Array): void;
-  copy_to_js(ptr: number, size: number): Float32Array;
-}
-
-export interface MemoryPool {
-  new (buffer_size: number, pool_size: number): MemoryPool;
-  acquire(): number | undefined;
-  release(index: number): void;
-  write_to_buffer(index: number, data: Float32Array, offset: number): boolean;
-  read_from_buffer(index: number): Float32Array;
-  apply_operation_to_buffer(index: number, operation: string): boolean;
-  free(): void;
-}
-
-export interface Vec2 {
-  new (x: number, y: number): Vec2;
-  x: number;
-  y: number;
-}
-
-
-export interface SpatialGrid {
-  new (cell_size: number, canvas_width: number, canvas_height: number): SpatialGrid;
-  clear(): void;
-  add_object(id: number, x: number, y: number, radius: number, is_visible: boolean): void;
-  update_positions(positions: Float32Array, radii: Float32Array, visibilities: Uint8Array): void;
-  find_overlaps(overlap_factor: number): Float32Array;
-  get_stats(): Float32Array;
-  get_cell_occupancy(): Float32Array;
-  free?(): void;
-}
-
-export interface ParticlePool {
-  new (): ParticlePool;
-  allocate_block(count: number, system_id: number): number[] | undefined;
-  allocate(system_id: number): number | undefined;
-  free_block(indices: number[]): void;
-  free(index: number): void;
-  free_system(system_id: number): void;
-  get_free_count(): number;
-  get_allocated_count(): number;
-  get_total_capacity(): number;
-  is_allocated_to(index: number, system_id: number): boolean;
-}
-
-export interface NebulaSystem {
-  new (canvas_width: number, canvas_height: number): NebulaSystem;
-  update_canvas_size(width: number, height: number): void;
-  init_particles(pool: ParticlePool, count: number): boolean;
-  update(delta_time: number, pool: ParticlePool): void;
-  get_render_data(): Float32Array;
-  get_color_data(): Float32Array;
-  find_overlaps(overlap_threshold: number): Float32Array;
-  release(pool: ParticlePool): void;
-  get_active_count(): number;
 }
 
 export async function loadWASM(): Promise<WASMModule | null> {
@@ -169,10 +49,7 @@ export async function loadWASM(): Promise<WASMModule | null> {
       // Initialize WASM module - this returns the wasm exports
       const wasmExports = await wasm.default(wasmPath);
       
-      // Get memory from the wasm exports
-      let wasmMemory = wasmExports?.memory || null;
-      
-      // Store module functions
+      // Store module functions - only active functions
       wasmModule = {
         add: wasm.add,
         greet: wasm.greet,
@@ -180,25 +57,11 @@ export async function loadWASM(): Promise<WASMModule | null> {
         generate_star_positions: wasm.generate_star_positions,
         generate_star_colors: wasm.generate_star_colors,
         generate_star_sizes: wasm.generate_star_sizes,
-        calculate_star_effects: wasm.calculate_star_effects,
-        calculate_star_effects_into_buffers: wasm.calculate_star_effects_into_buffers,
         calculate_star_effects_arrays: wasm.calculate_star_effects_arrays,
-        calculate_twinkle_effects: wasm.calculate_twinkle_effects,
-        calculate_sparkle_effects: wasm.calculate_sparkle_effects,
-        calculate_rotation_delta: wasm.calculate_rotation_delta,
-        calculate_speed_multiplier: wasm.calculate_speed_multiplier,
+        calculate_star_effects_with_temporal_coherence: wasm.calculate_star_effects_with_temporal_coherence,
         // Camera frustum culling
-        cull_stars_by_frustum: wasm.cull_stars_by_frustum,
         get_visible_star_indices: wasm.get_visible_star_indices,
         cull_stars_by_frustum_simd: wasm.cull_stars_by_frustum_simd,
-        // Enhanced SIMD batch processing for LOD groups
-        calculate_star_effects_by_lod: wasm.calculate_star_effects_by_lod,
-        // Temporal coherence optimization
-        calculate_star_effects_with_temporal_coherence: wasm.calculate_star_effects_with_temporal_coherence,
-        get_stars_needing_update: wasm.get_stars_needing_update,
-        calculate_star_effects_temporal_simd: wasm.calculate_star_effects_temporal_simd,
-        // LOD distribution calculations
-        calculate_lod_distribution: wasm.calculate_lod_distribution,
         // Math utilities
         fast_sin: wasm.fast_sin,
         fast_cos: wasm.fast_cos,
@@ -206,34 +69,6 @@ export async function loadWASM(): Promise<WASMModule | null> {
         fast_cos_batch: wasm.fast_cos_batch,
         seed_random: wasm.seed_random,
         seed_random_batch: wasm.seed_random_batch,
-        // Bezier functions
-        precalculate_bezier_path: wasm.precalculate_bezier_path,
-        precalculate_bezier_paths_batch: wasm.precalculate_bezier_paths_batch,
-        interpolate_bezier_point: wasm.interpolate_bezier_point,
-        precalculate_cubic_bezier_path: wasm.precalculate_cubic_bezier_path,
-        calculate_bezier_length: wasm.calculate_bezier_length,
-        // Memory management
-        SharedBuffer: wasm.SharedBuffer,
-        DirectMemory: wasm.DirectMemory,
-        MemoryPool: wasm.MemoryPool,
-        batch_process_sin: wasm.batch_process_sin,
-        batch_process_cos: wasm.batch_process_cos,
-        batch_process_with_operation: wasm.batch_process_with_operation,
-        // Spatial indexing system
-        SpatialGrid: wasm.SpatialGrid,
-        // Particle pool and utilities
-        ParticlePool: wasm.ParticlePool,
-        PhysicsUtils: wasm.PhysicsUtils,
-        Force: wasm.Force,
-        FastRandom: wasm.FastRandom,
-        BatchTransfer: wasm.BatchTransfer,
-        TypedBatchTransfer: wasm.TypedBatchTransfer,
-        ViewTransfer: wasm.ViewTransfer,
-        NebulaSystem: wasm.NebulaSystem,
-        // Export WASM memory for direct access
-        __wbindgen_export_0: wasmExports?.__wbindgen_export_0,
-        __wbg_memory: wasmExports?.__wbg_memory,
-        memory: wasmMemory,
       };
       
     } catch (error) {
@@ -246,8 +81,7 @@ export async function loadWASM(): Promise<WASMModule | null> {
   return wasmModule;
 }
 
-
-// JavaScript fallback implementations
+// JavaScript fallback implementations - essential functions only
 export const jsFallbacks: WASMModule = {
   add: (a: number, b: number): number => {
         return a + b;
@@ -318,13 +152,6 @@ export const jsFallbacks: WASMModule = {
     }
     return sizes;
   },
-  calculate_star_effects: (): Float32Array => {
-    return new Float32Array(0);
-  },
-  calculate_star_effects_into_buffers: (positions_ptr: number, twinkles_ptr: number, sparkles_ptr: number, count: number, time: number): void => {
-    // This is a void function, so we don't need to return anything
-    // In the actual component, we'll fall back to the manual calculation
-  },
   calculate_star_effects_arrays: (positions: Float32Array, count: number, time: number): Float32Array => {
     const effects = new Float32Array(count * 2);
     
@@ -346,88 +173,43 @@ export const jsFallbacks: WASMModule = {
     
     return effects;
   },
-  calculate_twinkle_effects: (): Float32Array => {
-    return new Float32Array(0);
-  },
-  calculate_sparkle_effects: (): Float32Array => {
-    return new Float32Array(0);
-  },
-  calculate_rotation_delta: (base_speed_x: number, base_speed_y: number, speed_multiplier: number, delta_time: number): Float32Array => {
-    return new Float32Array([
-      base_speed_x * speed_multiplier * delta_time,
-      base_speed_y * speed_multiplier * delta_time
-    ]);
-  },
-  calculate_speed_multiplier: (is_moving: boolean, click_time: number, current_time: number, current_multiplier: number): number => {
-    let speed_multiplier = 1.0;
+  calculate_star_effects_with_temporal_coherence: (
+    positions: Float32Array,
+    previous_twinkles: Float32Array,
+    previous_sparkles: Float32Array,
+    count: number,
+    time: number,
+    threshold: number
+  ): Float32Array => {
+    // Result format: [need_update_flag, twinkle, sparkle] triplets
+    const results = new Float32Array(count * 3);
     
-    // Apply movement boost
-    if (is_moving) {
-      speed_multiplier *= 4.5;
-    }
-    
-    // Apply click boost with decay
-    const time_since_click = current_time - click_time;
-    if (time_since_click < 1200) {
-      const click_decay = 1 - time_since_click / 1200;
-      const click_boost = 1 + 4.3 * click_decay;
-      speed_multiplier *= click_boost;
-    }
-    
-    // Apply smoothing (lerp with factor 0.2)
-    return current_multiplier + (speed_multiplier - current_multiplier) * 0.2;
-  },
-  // Camera frustum culling fallbacks
-  cull_stars_by_frustum: (positions: Float32Array, count: number, camera_matrix: Float32Array, fov: number, aspect_ratio: number, near: number, far: number): Uint8Array => {
-    
-    if (camera_matrix.length !== 16) {
-      return new Uint8Array(count).fill(1);
-    }
-    
-    const visibility_mask = new Uint8Array(count);
-    
-    // Extract frustum planes from view projection matrix
-    const m = camera_matrix;
-    const planes = [
-      [m[3] + m[0], m[7] + m[4], m[11] + m[8], m[15] + m[12]], // Left
-      [m[3] - m[0], m[7] - m[4], m[11] - m[8], m[15] - m[12]], // Right
-      [m[3] + m[1], m[7] + m[5], m[11] + m[9], m[15] + m[13]], // Bottom
-      [m[3] - m[1], m[7] - m[5], m[11] - m[9], m[15] - m[13]], // Top
-      [m[3] + m[2], m[7] + m[6], m[11] + m[10], m[15] + m[14]], // Near
-      [m[3] - m[2], m[7] - m[6], m[11] - m[10], m[15] - m[14]], // Far
-    ];
-    
-    // Normalize planes
-    const normalized_planes = planes.map(plane => {
-      const length = Math.sqrt(plane[0] * plane[0] + plane[1] * plane[1] + plane[2] * plane[2]);
-      return length > 0 ? plane.map(v => v / length) : plane;
-    });
-    
-    // Check each star
     for (let i = 0; i < count; i++) {
       const i3 = i * 3;
       const x = positions[i3];
       const y = positions[i3 + 1];
-      const z = positions[i3 + 2];
       
-      let inside = true;
+      // Calculate new effects
+      const twinkle_base = Math.sin(time * 3.0 + x * 10.0 + y * 10.0) * 0.3 + 0.7;
+      const sparkle_phase = Math.sin(time * 15.0 + x * 20.0 + y * 30.0);
+      const sparkle = sparkle_phase > 0.98 ? (sparkle_phase - 0.98) / 0.02 : 0.0;
+      const twinkle = twinkle_base + sparkle;
       
-      for (const plane of normalized_planes) {
-        const distance = plane[0] * x + plane[1] * y + plane[2] * z + plane[3];
-        
-        if (distance < -2.0) {
-          inside = false;
-          break;
-        }
-      }
+      // Check if update is needed
+      const twinkle_diff = Math.abs(twinkle - previous_twinkles[i]);
+      const sparkle_diff = Math.abs(sparkle - previous_sparkles[i]);
       
-      visibility_mask[i] = inside ? 1 : 0;
+      const needs_update = twinkle_diff > threshold || sparkle_diff > threshold;
+      
+      const idx = i * 3;
+      results[idx] = needs_update ? 1.0 : 0.0;
+      results[idx + 1] = twinkle;
+      results[idx + 2] = sparkle;
     }
     
-    return visibility_mask;
+    return results;
   },
   get_visible_star_indices: (positions: Float32Array, count: number, camera_matrix: Float32Array, margin: number): Uint32Array => {
-    
     if (camera_matrix.length !== 16) {
       return new Uint32Array(Array.from({length: count}, (_, i) => i));
     }
@@ -477,209 +259,6 @@ export const jsFallbacks: WASMModule = {
     return new Uint32Array(visible_indices);
   },
   cull_stars_by_frustum_simd: undefined, // No SIMD fallback
-  // Enhanced SIMD batch processing for LOD groups
-  calculate_star_effects_by_lod: (
-    near_positions: Float32Array,
-    near_count: number,
-    medium_positions: Float32Array,
-    medium_count: number,
-    far_positions: Float32Array,
-    far_count: number,
-    time: number,
-    quality_tier: number
-  ): Float32Array => {
-    
-    const total_count = near_count + medium_count + far_count;
-    const effects = new Float32Array(total_count * 2);
-    let offset = 0;
-    
-    // Helper function for full effects
-    const processFullEffects = (positions: Float32Array, count: number, startOffset: number) => {
-      for (let i = 0; i < count; i++) {
-        const i3 = i * 3;
-        const x = positions[i3];
-        const y = positions[i3 + 1];
-        
-        const twinkle_base = Math.sin(time * 3.0 + x * 10.0 + y * 10.0) * 0.3 + 0.7;
-        const sparkle_phase = Math.sin(time * 15.0 + x * 20.0 + y * 30.0);
-        const sparkle = sparkle_phase > 0.98 ? (sparkle_phase - 0.98) / 0.02 : 0.0;
-        
-        const idx = (startOffset + i) * 2;
-        effects[idx] = twinkle_base + sparkle;
-        effects[idx + 1] = sparkle;
-      }
-    };
-    
-    // Helper function for simple effects
-    const processSimpleEffects = (positions: Float32Array, count: number, startOffset: number) => {
-      for (let i = 0; i < count; i++) {
-        const i3 = i * 3;
-        const x = positions[i3];
-        const y = positions[i3 + 1];
-        
-        const twinkle = Math.sin(time * 2.0 + x * 5.0 + y * 5.0) * 0.2 + 0.8;
-        
-        const idx = (startOffset + i) * 2;
-        effects[idx] = twinkle;
-        effects[idx + 1] = 0.0;
-      }
-    };
-    
-    // Helper function for no effects
-    const processNoEffects = (count: number, startOffset: number) => {
-      for (let i = 0; i < count; i++) {
-        const idx = (startOffset + i) * 2;
-        effects[idx] = 1.0;
-        effects[idx + 1] = 0.0;
-      }
-    };
-    
-    // Process based on quality tier
-    switch (quality_tier) {
-      case 0: // Performance
-        processSimpleEffects(near_positions, near_count, offset);
-        offset += near_count;
-        processSimpleEffects(medium_positions, medium_count, offset);
-        offset += medium_count;
-        processNoEffects(far_count, offset);
-        break;
-      case 1: // Balanced
-        processFullEffects(near_positions, near_count, offset);
-        offset += near_count;
-        processSimpleEffects(medium_positions, medium_count, offset);
-        offset += medium_count;
-        processNoEffects(far_count, offset);
-        break;
-      default: // Ultra
-        processFullEffects(near_positions, near_count, offset);
-        offset += near_count;
-        processFullEffects(medium_positions, medium_count, offset);
-        offset += medium_count;
-        processSimpleEffects(far_positions, far_count, offset);
-        break;
-    }
-    
-    return effects;
-  },
-  // Temporal coherence optimization fallbacks
-  calculate_star_effects_with_temporal_coherence: (
-    positions: Float32Array,
-    previous_twinkles: Float32Array,
-    previous_sparkles: Float32Array,
-    count: number,
-    time: number,
-    threshold: number
-  ): Float32Array => {
-    
-    // Result format: [need_update_flag, twinkle, sparkle] triplets
-    const results = new Float32Array(count * 3);
-    
-    for (let i = 0; i < count; i++) {
-      const i3 = i * 3;
-      const x = positions[i3];
-      const y = positions[i3 + 1];
-      
-      // Calculate new effects
-      const twinkle_base = Math.sin(time * 3.0 + x * 10.0 + y * 10.0) * 0.3 + 0.7;
-      const sparkle_phase = Math.sin(time * 15.0 + x * 20.0 + y * 30.0);
-      const sparkle = sparkle_phase > 0.98 ? (sparkle_phase - 0.98) / 0.02 : 0.0;
-      const twinkle = twinkle_base + sparkle;
-      
-      // Check if update is needed
-      const twinkle_diff = Math.abs(twinkle - previous_twinkles[i]);
-      const sparkle_diff = Math.abs(sparkle - previous_sparkles[i]);
-      
-      const needs_update = twinkle_diff > threshold || sparkle_diff > threshold;
-      
-      const idx = i * 3;
-      results[idx] = needs_update ? 1.0 : 0.0;
-      results[idx + 1] = twinkle;
-      results[idx + 2] = sparkle;
-    }
-    
-    return results;
-  },
-  get_stars_needing_update: (
-    positions: Float32Array,
-    previous_twinkles: Float32Array,
-    previous_sparkles: Float32Array,
-    count: number,
-    time: number,
-    threshold: number
-  ): Uint32Array => {
-    
-    const indices: number[] = [];
-    
-    for (let i = 0; i < count; i++) {
-      const i3 = i * 3;
-      const x = positions[i3];
-      const y = positions[i3 + 1];
-      
-      // Calculate new effects
-      const twinkle_base = Math.sin(time * 3.0 + x * 10.0 + y * 10.0) * 0.3 + 0.7;
-      const sparkle_phase = Math.sin(time * 15.0 + x * 20.0 + y * 30.0);
-      const sparkle = sparkle_phase > 0.98 ? (sparkle_phase - 0.98) / 0.02 : 0.0;
-      const twinkle = twinkle_base + sparkle;
-      
-      // Check if update is needed
-      const twinkle_diff = Math.abs(twinkle - previous_twinkles[i]);
-      const sparkle_diff = Math.abs(sparkle - previous_sparkles[i]);
-      
-      if (twinkle_diff > threshold || sparkle_diff > threshold) {
-        indices.push(i);
-      }
-    }
-    
-    return new Uint32Array(indices);
-  },
-  calculate_star_effects_temporal_simd: undefined, // No SIMD fallback
-  // LOD distribution calculations
-  calculate_lod_distribution: (total_count: number, quality_tier: number): Uint32Array => {
-    
-    // Distribution ratios for each quality tier
-    let near_ratio: number;
-    let medium_ratio: number;
-    
-    switch (quality_tier) {
-      case 0: // Performance
-        near_ratio = 0.1;
-        medium_ratio = 0.3;
-        break;
-      case 1: // Balanced
-        near_ratio = 0.15;
-        medium_ratio = 0.35;
-        break;
-      default: // Ultra
-        near_ratio = 0.2;
-        medium_ratio = 0.4;
-        break;
-    }
-    
-    const near_count = Math.floor(total_count * near_ratio);
-    const medium_count = Math.floor(total_count * medium_ratio);
-    const far_count = total_count - near_count - medium_count;
-    
-    return new Uint32Array([near_count, medium_count, far_count]);
-  },
-  // Frame rate calculation
-  calculate_fps: (frame_count: number, current_time: number, last_time: number): Float32Array => {
-    
-    // Check if we should calculate FPS (every 30 frames)
-    if (frame_count % 30 === 0) {
-      // Calculate FPS: 30 frames / time_elapsed (in seconds)
-      const time_elapsed = current_time - last_time;
-      const fps = time_elapsed > 0 ? 30000 / time_elapsed : 60.0;
-      
-      // Return [fps, 1.0 (should update), current_time split into two f32s]
-      const time_high = Math.floor(current_time / 1000);
-      const time_low = current_time - (time_high * 1000);
-      
-      return new Float32Array([fps, 1.0, time_high, time_low]);
-    } else {
-      // Don't update FPS
-      return new Float32Array([0.0, 0.0, 0.0, 0.0]);
-    }
-  },
   // Math utilities
   fast_sin: (x: number): number => Math.sin(x),
   fast_cos: (x: number): number => Math.cos(x),
@@ -709,176 +288,6 @@ export const jsFallbacks: WASMModule = {
     }
     return result;
   },
-  // Bezier fallbacks
-  precalculate_bezier_path: (start_x: number, start_y: number, control_x: number, control_y: number, end_x: number, end_y: number, segments: number): Float32Array => {
-    const points = new Float32Array((segments + 1) * 2);
-    
-    for (let i = 0; i <= segments; i++) {
-      const t = i / segments;
-      const oneMinusT = 1 - t;
-      
-      // Quadratic Bezier formula
-      const x = oneMinusT * oneMinusT * start_x + 
-                2 * oneMinusT * t * control_x + 
-                t * t * end_x;
-      
-      const y = oneMinusT * oneMinusT * start_y + 
-                2 * oneMinusT * t * control_y + 
-                t * t * end_y;
-      
-      points[i * 2] = x;
-      points[i * 2 + 1] = y;
-    }
-    
-    return points;
-  },
-  precalculate_bezier_paths_batch: (paths_data: Float32Array, segments: number): Float32Array => {
-    const pathCount = paths_data.length / 6;
-    const pointsPerPath = (segments + 1) * 2;
-    const allPoints = new Float32Array(pathCount * pointsPerPath);
-    
-    for (let pathIdx = 0; pathIdx < pathCount; pathIdx++) {
-      const baseIdx = pathIdx * 6;
-      const start_x = paths_data[baseIdx];
-      const start_y = paths_data[baseIdx + 1];
-      const control_x = paths_data[baseIdx + 2];
-      const control_y = paths_data[baseIdx + 3];
-      const end_x = paths_data[baseIdx + 4];
-      const end_y = paths_data[baseIdx + 5];
-      
-      for (let i = 0; i <= segments; i++) {
-        const t = i / segments;
-        const oneMinusT = 1 - t;
-        
-        const x = oneMinusT * oneMinusT * start_x + 
-                  2 * oneMinusT * t * control_x + 
-                  t * t * end_x;
-        
-        const y = oneMinusT * oneMinusT * start_y + 
-                  2 * oneMinusT * t * control_y + 
-                  t * t * end_y;
-        
-        const outputIdx = pathIdx * pointsPerPath + i * 2;
-        allPoints[outputIdx] = x;
-        allPoints[outputIdx + 1] = y;
-      }
-    }
-    
-    return allPoints;
-  },
-  interpolate_bezier_point: (points: Float32Array, t: number): Float32Array => {
-    const pointCount = points.length / 2;
-    if (pointCount === 0) {
-      return new Float32Array([0, 0]);
-    }
-    
-    const index = Math.floor(t * (pointCount - 1));
-    const localT = (t * (pointCount - 1)) % 1;
-    
-    if (index >= pointCount - 1) {
-      const lastIdx = (pointCount - 1) * 2;
-      return new Float32Array([points[lastIdx], points[lastIdx + 1]]);
-    }
-    
-    const p1Idx = index * 2;
-    const p2Idx = (index + 1) * 2;
-    
-    return new Float32Array([
-      points[p1Idx] + (points[p2Idx] - points[p1Idx]) * localT,
-      points[p1Idx + 1] + (points[p2Idx + 1] - points[p1Idx + 1]) * localT
-    ]);
-  },
-  precalculate_cubic_bezier_path: (p0x: number, p0y: number, p1x: number, p1y: number, p2x: number, p2y: number, p3x: number, p3y: number, segments: number): Float32Array => {
-    const points = new Float32Array((segments + 1) * 2);
-    
-    for (let i = 0; i <= segments; i++) {
-      const t = i / segments;
-      const oneMinusT = 1 - t;
-      const oneMinusTSq = oneMinusT * oneMinusT;
-      const oneMinusTCube = oneMinusTSq * oneMinusT;
-      const tSq = t * t;
-      const tCube = tSq * t;
-      
-      // Cubic Bezier formula
-      const x = oneMinusTCube * p0x + 
-                3 * oneMinusTSq * t * p1x + 
-                3 * oneMinusT * tSq * p2x + 
-                tCube * p3x;
-      
-      const y = oneMinusTCube * p0y + 
-                3 * oneMinusTSq * t * p1y + 
-                3 * oneMinusT * tSq * p2y + 
-                tCube * p3y;
-      
-      points[i * 2] = x;
-      points[i * 2 + 1] = y;
-    }
-    
-    return points;
-  },
-  calculate_bezier_length: (points: Float32Array): number => {
-    const pointCount = points.length / 2;
-    if (pointCount < 2) {
-      return 0;
-    }
-    
-    let length = 0;
-    
-    for (let i = 1; i < pointCount; i++) {
-      const prevIdx = (i - 1) * 2;
-      const currIdx = i * 2;
-      
-      const dx = points[currIdx] - points[prevIdx];
-      const dy = points[currIdx + 1] - points[prevIdx + 1];
-      
-      length += Math.sqrt(dx * dx + dy * dy);
-    }
-    
-    return length;
-  },
-  // Memory management fallbacks (not implemented)
-  SharedBuffer: null as any,
-  DirectMemory: null as any,
-  MemoryPool: null as any,
-  batch_process_sin: (input: Float32Array): Float32Array => {
-    const result = new Float32Array(input.length);
-    for (let i = 0; i < input.length; i++) {
-      result[i] = Math.sin(input[i]);
-    }
-    return result;
-  },
-  batch_process_cos: (input: Float32Array): Float32Array => {
-    const result = new Float32Array(input.length);
-    for (let i = 0; i < input.length; i++) {
-      result[i] = Math.cos(input[i]);
-    }
-    return result;
-  },
-  batch_process_with_operation: (input: Float32Array, operation: string): Float32Array => {
-    const result = new Float32Array(input.length);
-    switch(operation) {
-      case 'sin':
-        for (let i = 0; i < input.length; i++) result[i] = Math.sin(input[i]);
-        break;
-      case 'cos':
-        for (let i = 0; i < input.length; i++) result[i] = Math.cos(input[i]);
-        break;
-      default:
-        return input;
-    }
-    return result;
-  },
-  // Spatial indexing system fallback (not implemented)
-  SpatialGrid: null as any,
-  // Particle pool and utilities fallbacks (not implemented)
-  ParticlePool: null as any,
-  PhysicsUtils: null as any,
-  Force: null as any,
-  FastRandom: null as any,
-  BatchTransfer: null as any,
-  TypedBatchTransfer: null as any,
-  ViewTransfer: null as any,
-  NebulaSystem: null as any,
 };
 
 // Unified API that automatically uses WASM or JS fallback
@@ -894,4 +303,3 @@ export function getWASMStatus(): { loaded: boolean; usingFallback: boolean } {
     usingFallback: isUsingFallback
   };
 }
-
